last minute clean ups
